pyproject.toml
[project]
name = "vilnius-demo-soniox"
version = "0.1.0"
description = "Vilnius savivaldybė LiveKit voice agent — Soniox STT + Soniox TTS variant"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.4",
    "livekit-plugins-silero>=1.4",
    "livekit-plugins-soniox>=1.4",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sioniox", "knowledge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
